--- constants/partners.js
@@ -1,26 +1,0 @@
-export const partners = [
-  {
-    name: 'SeedClub Ventures',
-    icon: '/../public/partners/seedclub-logo.svg',
-  },
-  {
-    name: 'Hashed',
-    icon: '../public/partners/hashed-logo.svg',
-  },
-  // {
-  //   name: "Terra Ecosystem",
-  //   icon: "",
-  // },
-  {
-    name: 'LexDAO',
-    icon: '/../public/partners/lexdao.png',
-  },
-  {
-    name: 'LexPunk Army',
-    icon: '/../public/partners/lexpunk.png',
-  },
-  // {
-  //   name: "Twali",
-  //   icon: "",
-  // },
-]
