--- constants/tweets.js
@@ -1,11 +1,0 @@
-export const tweets = [
-  { from: 'Engagemint_eth', tweetId: '1500881587004755974' },
-  { from: 'rhea_sg', tweetId: '1498668739423870979' },
-  { from: 'gcattnip', tweetId: '1496355189485359106' },
-  { from: 'lion917_eth', tweetId: '1495730950529249284' },
-  { from: 'exiledsurfer', tweetId: '1498562721150902276' },
-  { from: 'seedclubhq', tweetId: '1490757832076972032' },
-  { from: 'arjunmahadevan', tweetId: '1495262699852374018' },
-  { from: 'multisHQ', tweetId: '1502313425174609920' },
-  { from: 'cokiehasiotis', tweetId: '1499788936985653251' },
-]
