+++ package.json
@@ -19,8 +19,11 @@
     "react-dom": "17.0.2"
   },
   "devDependencies": {
+    "@types/node": "^18.0.0",
+    "@types/react": "18.0.1",
     "eslint": "8.11.0",
     "eslint-config-next": "12.1.0",
-    "prettier": "^2.6.2"
+    "prettier": "^2.6.2",
+    "typescript": "^4.7.4"
   }
 }
