+++ pages/_app.tsx
@@ -1,6 +1,7 @@
+import { AppProps } from 'next/app'
 import globalStyles from '../styles/globalStyles'
 
-function MyApp({ Component, pageProps }) {
+function MyApp({ Component, pageProps }: AppProps) {
   globalStyles()
   return <Component {...pageProps} />
 }
