+++ styles/layout.js
@@ -1,0 +1,7 @@
+import { styled } from './stitches.config'
+
+const Layout = styled('div', {
+  minHeight: '100vh',
+})
+
+export default Layout
