+++ styles/stream.js
@@ -1,0 +1,14 @@
+import { styled } from './stitches.config'
+
+const PartnerStream = styled('div', {
+  display: 'flex',
+  flexDirection: 'row',
+  justifyContent: 'center',
+  alignItems: 'center',
+  gap: '1rem',
+  zIndex: '0',
+  marginTop: '3rem',
+  pointerEvents: 'none',
+})
+
+export default PartnerStream
